+++ src/main/java/com/besson/arknights/painting/ModPaintingVariants.java
@@ -68,6 +68,7 @@ public class ModPaintingVariants {
     public static final PaintingVariant FRUITY_ADVENTURE3 = registerPaintingVariant("fruity_adventure3", new PaintingVariant(16, 16));
     public static final PaintingVariant FAMOUS_PAINTING_COPY = registerPaintingVariant("famous_painting_copy", new PaintingVariant(16, 16));
     public static final PaintingVariant FAMOUS_PAINTING_REPLICA = registerPaintingVariant("famous_painting_replica", new PaintingVariant(16, 16));
+    public static final PaintingVariant LOVE_AND_SUNFLOWERS = registerPaintingVariant("love_and_sunflowers", new PaintingVariant(16, 16));
 
     private static PaintingVariant registerPaintingVariant(String id, PaintingVariant variant) {
         return Registry.register(Registries.PAINTING_VARIANT, new Identifier(ArknightsFurniture.MOD_ID, id), variant);